downloads and batch processing;  if file exists, skip with a message

+++ pdf_slidedeck2book.sh
@@ -12,6 +12,7 @@ DPI=300
 BOOK_WIDTH=11
 BOOK_HEIGHT=8.5
 CLEANUP=1
+FORCE=0
 args=()
 VERBOSE=false
 
@@ -24,11 +25,12 @@ function usage
   echo "  $scriptname <file>        (list of pdf files to process)"
   echo "  $scriptname --help        (this help)"
   echo "  $scriptname --verbose     (output verbose information)"
-  echo "  $scriptname --margin      (default: ${MARGIN_SIZE})"
-  echo "  $scriptname --dpi         (default: ${DPI})"
-  echo "  $scriptname --width       (default: ${BOOK_WIDTH})"
-  echo "  $scriptname --height      (default: ${BOOK_HEIGHT})"
-  echo "  $scriptname --cleanup     (default: ${CLEANUP})"
+  echo "  $scriptname --margin      (default: --margin ${MARGIN_SIZE})"
+  echo "  $scriptname --dpi         (default: --dpi ${DPI})"
+  echo "  $scriptname --width       (default: --width ${BOOK_WIDTH})"
+  echo "  $scriptname --height      (default: --height ${BOOK_HEIGHT})"
+  echo "  $scriptname --cleanup     (default: --cleanup ${CLEANUP})"
+  echo "  $scriptname --force       (overwrite file(s), default: --force ${FORCE})"
   echo ""
 }
 ARGC=$#
@@ -74,6 +76,12 @@ for ((i = 0; i < ARGC; i++)); do
     $VERBOSE && echo "Parsing Args: Changing cleanup to $CLEANUP"
     continue
   fi
+  if [[ $ARGC -ge 1 && ${ARGV[$i]} == "--force" ]]; then
+    ((i+=1))
+    FORCE=${ARGV[$i]}
+    $VERBOSE && echo "Parsing Args: Changing force to $FORCE"
+    continue
+  fi
   if [[ $ARGC -ge 1 && ${ARGV[$i]:0:2} == "--" ]]; then
     echo "Unknown option ${ARGV[$i]}"
     exit -1
@@ -108,11 +116,14 @@ function filepath_ext { local file=$1; echo "${file##*.}"; }
 function cleanup()
 {
   if [ "$CLEANUP" == 1 ]; then
-    echo "Removing:"
-    echo " - $OUTDIR1"
-    rm -rf "$OUTDIR1"
-    echo " - $OUTDIR2"
-    rm -rf "$OUTDIR2"
+    if [ -d "$OUTDIR1" ]; then
+      echo "Removing: $OUTDIR1"
+      rm -rf "$OUTDIR1"
+    fi
+    if [ -d "$OUTDIR2" ]; then
+      echo "Removing: $OUTDIR2"
+      rm -rf "$OUTDIR2"
+    fi
   fi
 }
 function ctrl_c()
@@ -213,48 +224,81 @@ function addMarginToPNGs {
 
 
 # do the work:
-for INPUTFILE in "${args[@]}"; do
-  INPUTFILEPATH=`filepath_path "$INPUTFILE"`
-  INPUTFILENAME=`filepath_name "$INPUTFILE"`
-  OUTDIR1=`mktemp -d -t "$INPUTFILENAME-1"`
-  OUTDIR2=`mktemp -d -t "$INPUTFILENAME-2"`
-  echo "----------"
-  echo "processing $INPUTFILE"
+function process {
+  local DECO_BEGIN="--=::["
+  local DECO_END="]::=--"
+  local DECO_JOB="--==========::==========--"
+  local DECO_STEP="----------"
 
-  ######################################
-  # get dpi of pdf:
-  ######################################
-  echo "dpi of input pdf"
-  magick identify -format "%w x %h %x x %y\n" "$INPUTFILE"
+  for INPUTFILE in "${args[@]}"; do
+    local INPUTFILENAME=`filepath_name "$INPUTFILE"`
+    local OUTPATH="./$INPUTFILENAME-book.pdf"
 
-  ######################################
-  # CONVERT PDF TO PNGs
-  ######################################
-  echo "CONVERT PDF TO PNGs..."
-  pdf2pngs "$INPUTFILE" "$OUTDIR1" "$DPI"
+    # force remove previous file if present
+    if [ $FORCE -eq 1 ]; then
+      rm -f "${OUTPATH}"
+    fi
+    # skip job if output file already exists
+    if [ -f "${OUTPATH}" ]; then
+      echo "Skipping: \"$INPUTFILE\" -> \"$OUTPATH\" (exists)"
+      continue
+    fi
+
+    # Start job
+    echo "${DECO_JOB}"
+    echo "${DECO_BEGIN}Processing \"$INPUTFILE\"${DECO_END}"
+
+    # use global vars for OUTDIRX, so cleanup works
+    OUTDIR1=`mktemp -d -t "$INPUTFILENAME-1"`
+    OUTDIR2=`mktemp -d -t "$INPUTFILENAME-2"`
+
+    ######################################
+    # get dpi of pdf:
+    ######################################
+    #echo "${DECO_STEP}"
+    echo "${DECO_BEGIN}dpi of input pdf${DECO_END}"
+    magick identify -monitor -format "%w x %h %x x %y\n" "$INPUTFILE"
+
+    ######################################
+    # CONVERT PDF TO PNGs
+    ######################################
+    #echo "${DECO_STEP}"
+    echo "${DECO_BEGIN}CONVERT PDF PAGES TO PNG(s)...${DECO_END}"
+    pdf2pngs "$INPUTFILE" "$OUTDIR1" "$DPI"
 
 
-  ######################################
-  # ADD MARGINS TO PNGs
-  ######################################
-  echo "ADD MARGINS TO PNGs..."
-  addMarginToPNGs "$OUTDIR1" "$OUTDIR2" "$((DPI/2))" `python -c "print('{0:0.4f}'.format($BOOK_HEIGHT / $BOOK_WIDTH))"`
+    ######################################
+    # ADD MARGINS TO PNGs
+    ######################################
+    #echo "${DECO_STEP}"
+    echo "${DECO_BEGIN}ADD MARGINS TO PNG(s)...${DECO_END}"
+    addMarginToPNGs "$OUTDIR1" "$OUTDIR2" "$((DPI/2))" `python -c "print('{0:0.4f}'.format($BOOK_HEIGHT / $BOOK_WIDTH))"`
 
+    ######################################
+    # CONVERT PNGs to PDF
+    ######################################
+    #echo "${DECO_STEP}"
+    echo "${DECO_BEGIN}CONVERT PNGs to PDF...${DECO_END}"
+    echo " - silent while loading PNGs"
+    echo " - outputs % for each generated PDF page"
+    magick convert -monitor "$OUTDIR2/"*.png -alpha off -set units $DPI -quality 60 -compress jpeg "${OUTPATH}"
 
-  #####################################
+    # comparison of [jpeg, lzw, rle, zip] compression schemes:
+    # -rw-r--r--  1 kevinmeinert  staff    81861538 Apr  1 09:09 grow-jpeg.pdf
+    # -rw-r--r--  1 kevinmeinert  staff  1922920649 Apr  1 09:19 grow-lzw.pdf
+    # -rw-r--r--  1 kevinmeinert  staff  2249930760 Apr  1 09:22 grow-rle.pdf
+    # -rw-r--r--  1 kevinmeinert  staff  1190841219 Apr  1 09:13 grow-zip.pdf
 
-  ######################################
-  # CONVERT PNGs to PDF
-  ######################################
-  echo "CONVERT PNGs to PDF..."
-  rm "$INPUTFILEPATH/${INPUTFILENAME}-margins.pdf"
-  # -rw-r--r--  1 kevinmeinert  staff    81861538 Apr  1 09:09 grow-jpeg.pdf
-  # -rw-r--r--  1 kevinmeinert  staff  1922920649 Apr  1 09:19 grow-lzw.pdf
-  # -rw-r--r--  1 kevinmeinert  staff  2249930760 Apr  1 09:22 grow-rle.pdf
-  # -rw-r--r--  1 kevinmeinert  staff  1190841219 Apr  1 09:13 grow-zip.pdf
-  magick convert -monitor "$OUTDIR2/"*.png -alpha off -set units $DPI -quality 60 -compress jpeg "$INPUTFILEPATH/${INPUTFILENAME}-margins.pdf"
+    #####################################
+    # remove tempdirs
+    cleanup
 
+    #####################################
+    # open the new pdf
+    echo "${DECO_BEGIN}Opening '${OUTPATH}'${DECO_END}"
+    open "${OUTPATH}"
+  done
+}
 
-  echo "Opening '$INPUTFILEPATH/${INPUTFILENAME}-margins.pdf'"
-  open "$INPUTFILEPATH/${INPUTFILENAME}-margins.pdf"
-done
+process
+
